+++ block/fetcher/rpc.go
@@ -34,18 +34,23 @@ var GetBlockOpts = &rpc.GetBlockOpts{
 type fetchBlock func(ctx context.Context, requestedSlot uint64) (slot uint64, out *rpc.GetBlockResult, err error)
 
 type RPCFetcher struct {
+	rpcClients               *firecoreRPC.Clients[*rpc.Client]
+	optimizeForSingleTarget  bool
 	latestConfirmedSlot      uint64
 	latestFinalizedSlot      uint64
 	latestBlockRetryInterval time.Duration
 	fetchInterval            time.Duration
 	lastFetchAt              time.Time
+	isMainnet                bool
 	logger                   *zap.Logger
 }
 
-func NewRPC(fetchInterval time.Duration, latestBlockRetryInterval time.Duration, logger *zap.Logger) *RPCFetcher {
+func NewRPC(fetchInterval time.Duration, latestBlockRetryInterval time.Duration, optimizeForSingleTarget bool, isMainnet bool, logger *zap.Logger) *RPCFetcher {
 	f := &RPCFetcher{
 		fetchInterval:            fetchInterval,
+		optimizeForSingleTarget:  optimizeForSingleTarget,
 		latestBlockRetryInterval: latestBlockRetryInterval,
+		isMainnet:                isMainnet,
 		logger:                   logger,
 	}
 	return f
@@ -57,8 +62,8 @@ func (f *RPCFetcher) IsBlockAvailable(requestedSlot uint64) bool {
 }
 
 func (f *RPCFetcher) Fetch(ctx context.Context, client *rpc.Client, requestedSlot uint64) (b *pbbstream.Block, skipped bool, err error) {
-	//THIS IS A FKG Ugly hack!
-	if requestedSlot >= 13334464 && requestedSlot <= 13334475 {
+	if f.isMainnet && requestedSlot >= 13334464 && requestedSlot <= 13334475 {
+		// know issue fetching these blocks on mainnet, ugly but works
 		return nil, true, nil
 	}